Ecommerce
An Ecommerce company is trying to decide whether to focus their efforts on their mobile app experience or their website. The given data is Email, Address, Avatar, Avg. Session Length, Time on App, Time on Website, Length of Membership, Yearly Amount Spent. and we are trying different parameters to measure the cases in which the Yearly Amount Spent is increased.