Add files via upload
In this project we will be working with a fake advertising data set, indicating whether or not a particular internet user clicked on an Advertisement. we will try to create a model that will predict whether or not they will click on an ad-based of the features of that user.